REST frameworks/ReservaProject/requirements.txt
asgiref==3.10.0
Django==5.2.7
djangorestframework==3.16.1
pillow==11.3.0
py==1.11.0
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.10
django-encrypted-model-fields==0.6.5
django-cryptography==1.1
django-cors-headers==4.6.0
django-filter==24.3
cryptography==46.0.3

# WebSockets (Channels)
channels==4.0.0
channels-redis==4.2.0

# Dependencias para producción (Railway/Heroku)
gunicorn==23.0.0
whitenoise==6.9.0
dj-database-url==2.3.0
